rename discordUser to Discord schema
