+++ src/app/homebrew-sharing/tome-share/tome-share.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {Meta} from '@angular/platform-browser';
 import {ActivatedRoute} from '@angular/router';
 import {environment} from '../../../environments/environment';
-import {DiscordUser} from '../../schemas/DiscordUser';
+import {DiscordUser} from '../../schemas/Discord';
 import {parseLevel, parseSchool, Spell, Tome} from '../../schemas/homebrew/Spells';
 import {DiscordService} from '../../shared/discord.service';
 import {HomebrewSharingService} from '../homebrew-sharing.service';
